DokanFileInfo: no need to dispose twice the SafeFileHandle m using get's lowered to a tryfinally block. DokanFileInfo: More verbose and explicit way of getting a GCHandle and an IntPtr from it. DokanInstance: Leverage language expressiveness to make clearer what should not be changed, also expose a property to show if it is disposed. Mirror Notify class is now disposable and check if DokanInstance has been already disposed before making any call. Mirror and RegistryFs now handle the CancelKeyPress gracefully.
