+++ DokanNet/DokanInstance.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Runtime.ConstrainedExecution;
-using System.Runtime.InteropServices;
-using System.Text;
 using DokanNet.Native;
 
 namespace DokanNet
@@ -13,31 +9,52 @@ namespace DokanNet
     /// </summary>
     public class DokanInstance : IDisposable
     {
-        internal NativeStructWrapper<DOKAN_OPTIONS> DokanOptions;
-        internal NativeStructWrapper<DOKAN_OPERATIONS> DokanOperations;
-        internal DokanHandle DokanHandle;
-        private bool disposedValue;
+        internal NativeStructWrapper<DOKAN_OPTIONS> DokanOptions { get; private set; }
+        internal NativeStructWrapper<DOKAN_OPERATIONS> DokanOperations { get; private set; }
+        internal DokanHandle DokanHandle { get; private set; }
+        private readonly object _disposeLock;
+        private bool _disposed = false;
+        public bool IsDisposed
+        {
+            get { lock (_disposeLock) return _disposed; }
+        }
+
+        internal DokanInstance(DOKAN_OPTIONS options, DOKAN_OPERATIONS operations)
+        {
+            DokanOptions = NativeStructWrapper.Wrap(options);
+            DokanOperations = NativeStructWrapper.Wrap(operations);
+            _disposeLock = new object();
+            var status = NativeMethods.DokanCreateFileSystem(DokanOptions, DokanOperations, out var handle);
+            if (status != DokanStatus.Success)
+            {
+                throw new DokanException(status);
+            }
+            DokanHandle = handle;
+        }
 
         protected virtual void Dispose(bool disposing)
         {
-            if (!disposedValue)
+            lock (_disposeLock)
             {
-                if (disposing)
+                if (!_disposed)
                 {
-                    // Dispose managed state (managed objects)
-                    DokanHandle?.Dispose();     // This calls DokanCloseHandle and waits for dismount
-                    DokanOptions?.Dispose();    // After that, it is safe to free unmanaged memory
-                    DokanOperations?.Dispose();
+                    if (disposing)
+                    {
+                        // Dispose managed state (managed objects)
+                        DokanHandle?.Dispose();     // This calls DokanCloseHandle and waits for dismount
+                        DokanOptions?.Dispose();    // After that, it is safe to free unmanaged memory
+                        DokanOperations?.Dispose();
+                    }
+
+                    // Free unmanaged resources (unmanaged objects) and override finalizer
+
+                    // Set fields to null
+                    DokanOptions = null;
+                    DokanOperations = null;
+                    DokanHandle = null;
+
+                    _disposed = true;
                 }
-
-                // Free unmanaged resources (unmanaged objects) and override finalizer
-
-                // Set fields to null
-                DokanOptions = null;
-                DokanOperations = null;
-                DokanHandle = null;
-
-                disposedValue = true;
             }
         }
